Separate initial views to a util class
