+++ src/main/kotlin/com/mituuz/fuzzier/Fuzzier.kt
@@ -28,7 +28,6 @@ import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.editor.Document
 import com.intellij.openapi.editor.EditorFactory
 import com.intellij.openapi.fileEditor.FileEditorManager
-import com.intellij.openapi.fileEditor.impl.EditorHistoryManager
 import com.intellij.openapi.module.ModuleManager
 import com.intellij.openapi.progress.ProgressIndicator
 import com.intellij.openapi.progress.ProgressManager
@@ -48,10 +47,12 @@ import com.intellij.openapi.wm.WindowManager
 import com.mituuz.fuzzier.components.FuzzyFinderComponent
 import com.mituuz.fuzzier.entities.FuzzyMatchContainer
 import com.mituuz.fuzzier.entities.StringEvaluator
-import com.mituuz.fuzzier.settings.FuzzierSettingsService
 import com.mituuz.fuzzier.settings.FuzzierSettingsService.RecentFilesMode.NONE
+import com.mituuz.fuzzier.settings.FuzzierSettingsService.RecentFilesMode.RECENTLY_SEARCHED_FILES
+import com.mituuz.fuzzier.settings.FuzzierSettingsService.RecentFilesMode.RECENT_PROJECT_FILES
 import com.mituuz.fuzzier.util.FuzzierUtil
 import com.mituuz.fuzzier.util.FuzzierUtil.Companion.createDimensionKey
+import com.mituuz.fuzzier.util.InitialviewHandler
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
@@ -68,6 +69,7 @@ open class Fuzzier : FuzzyAction() {
     private var defaultDoc: Document? = null
     open var title: String = "Fuzzy Search"
     private val fuzzyDimensionKey: String = "FuzzySearchPopup"
+
     // Used by FuzzierVCS to check if files are tracked by the VCS
     protected var changeListManager: ChangeListManager? = null
 
@@ -136,8 +138,13 @@ open class Fuzzier : FuzzyAction() {
     private fun createInitialView(project: Project) {
         ApplicationManager.getApplication().executeOnPooledThread {
             val listModel = when (fuzzierSettingsService.state.recentFilesMode) {
-                FuzzierSettingsService.RecentFilesMode.RECENT_PROJECT_FILES -> getRecentProjectFiles(project)
-                FuzzierSettingsService.RecentFilesMode.RECENTLY_SEARCHED_FILES -> getRecentlySearchedFiles()
+                RECENT_PROJECT_FILES -> InitialviewHandler.getRecentProjectFiles(
+                    project,
+                    fuzzierSettingsService,
+                    fuzzierUtil
+                )
+
+                RECENTLY_SEARCHED_FILES -> InitialviewHandler.getRecentlySearchedFiles(fuzzierSettingsService)
                 else -> {
                     DefaultListModel<FuzzyMatchContainer>()
                 }
@@ -152,50 +159,6 @@ open class Fuzzier : FuzzyAction() {
                 }
             }
         }
-    }
-
-    private fun getRecentProjectFiles(project: Project): DefaultListModel<FuzzyMatchContainer> {
-        val editorHistory = EditorHistoryManager.getInstance(project).fileList
-        val listModel = DefaultListModel<FuzzyMatchContainer>()
-        val limit = fuzzierSettingsService.state.fileListLimit
-
-        // Start from the end of editor history (most recent file)
-        var i = editorHistory.size - 1
-        while (i >= 0 && listModel.size() < limit) {
-            val file = editorHistory[i]
-            val filePathAndModule = fuzzierUtil.removeModulePath(file.path)
-            // Don't add files that do not have a module path in the project
-            if (filePathAndModule.second == "") {
-                i--
-                continue
-            }
-            val fuzzyMatchContainer =
-                FuzzyMatchContainer.createOrderedContainer(i, filePathAndModule.first, filePathAndModule.second, file.name)
-            listModel.addElement(fuzzyMatchContainer)
-            i--
-        }
-
-        return listModel
-    }
-
-    private fun getRecentlySearchedFiles(): DefaultListModel<FuzzyMatchContainer> {
-        var listModel = fuzzierSettingsService.state.recentlySearchedFiles
-
-        if (listModel == null) {
-            listModel = DefaultListModel<FuzzyMatchContainer>()
-            fuzzierSettingsService.state.recentlySearchedFiles = listModel
-        }
-
-        var i = 0
-        while (i < listModel.size) {
-            if (listModel.get(i) == null) {
-                listModel.remove(i)
-            } else {
-                i++
-            }
-        }
-
-        return listModel
     }
 
     override fun updateListContents(project: Project, searchString: String) {
@@ -239,7 +202,7 @@ open class Fuzzier : FuzzyAction() {
             }
         }
     }
-    
+
     private fun handleEmptySearchString(project: Project) {
         if (fuzzierSettingsService.state.recentFilesMode != NONE) {
             createInitialView(project)
@@ -250,7 +213,7 @@ open class Fuzzier : FuzzyAction() {
             }
         }
     }
-    
+
     private fun getStringEvaluator(): StringEvaluator {
         return StringEvaluator(
             fuzzierSettingsService.state.exclusionSet,
@@ -258,9 +221,11 @@ open class Fuzzier : FuzzyAction() {
             changeListManager
         )
     }
-    
-    private fun process(project: Project, stringEvaluator: StringEvaluator, searchString: String,
-                        listModel: DefaultListModel<FuzzyMatchContainer>, task: Future<*>?) {
+
+    private fun process(
+        project: Project, stringEvaluator: StringEvaluator, searchString: String,
+        listModel: DefaultListModel<FuzzyMatchContainer>, task: Future<*>?
+    ) {
         val moduleManager = ModuleManager.getInstance(project)
         if (fuzzierSettingsService.state.isProject) {
             processProject(project, stringEvaluator, searchString, listModel, task)
@@ -268,16 +233,20 @@ open class Fuzzier : FuzzyAction() {
             processModules(moduleManager, stringEvaluator, searchString, listModel, task)
         }
     }
-    
-    private fun processProject(project: Project, stringEvaluator: StringEvaluator,
-                               searchString: String, listModel: DefaultListModel<FuzzyMatchContainer>, task: Future<*>?) {
+
+    private fun processProject(
+        project: Project, stringEvaluator: StringEvaluator,
+        searchString: String, listModel: DefaultListModel<FuzzyMatchContainer>, task: Future<*>?
+    ) {
         val filesToIterate = ConcurrentHashMap.newKeySet<FuzzierUtil.IterationFile>()
         FuzzierUtil.fileIndexToIterationFile(filesToIterate, ProjectFileIndex.getInstance(project), project.name, task)
         processFiles(filesToIterate, stringEvaluator, listModel, searchString, task)
     }
 
-    private fun processModules(moduleManager: ModuleManager, stringEvaluator: StringEvaluator,
-                               searchString: String, listModel: DefaultListModel<FuzzyMatchContainer>, task: Future<*>?) {
+    private fun processModules(
+        moduleManager: ModuleManager, stringEvaluator: StringEvaluator,
+        searchString: String, listModel: DefaultListModel<FuzzyMatchContainer>, task: Future<*>?
+    ) {
         val filesToIterate = ConcurrentHashMap.newKeySet<FuzzierUtil.IterationFile>()
         for (module in moduleManager.modules) {
             FuzzierUtil.fileIndexToIterationFile(filesToIterate, module.rootManager.fileIndex, module.name, task)
